dev-requirements.txt
pyramid_debugtoolbar
pytest
webtest